move LastMessageData to its own file
