+++ ConsentViewController/Classes/Consents/SPGDPRConsent.swift
@@ -7,19 +7,6 @@
 
 import Foundation
 
-struct LastMessageData: Codable {
-    var id, categoryId, subCategoryId: Int
-    var partitionUUID: String
-}
-
-extension LastMessageData {
-    init(from metadata: MessageMetaData) {
-        id = Int(metadata.messageId) ?? 0
-        categoryId = metadata.categoryId.rawValue
-        subCategoryId = metadata.subCategoryId.rawValue
-        partitionUUID = metadata.messagePartitionUUID ?? ""
-    }
-}
 
 /// A dictionary in which the keys represent the Vendor Id
 public typealias SPGDPRVendorGrants = [GDPRVendorId: SPGDPRVendorGrant]
